+++ app/src/main/java/net/aiscope/gdd_app/ui/metadata/MetadataPresenter.kt
@@ -6,7 +6,7 @@ import net.aiscope.gdd_app.model.SampleMetadata
 import net.aiscope.gdd_app.model.SampleStatus
 import net.aiscope.gdd_app.network.RemoteStorage
 import net.aiscope.gdd_app.repository.SampleRepository
-import net.aiscope.gdd_app.ui.sample_completion.MetadataMapper
+import net.aiscope.gdd_app.ui.sample_completion.metadata.MetadataMapper
 import timber.log.Timber
 import javax.inject.Inject
