+++ app/src/main/java/net/aiscope/gdd_app/ui/sample_completion/FragmentAdapter.kt
@@ -4,6 +4,9 @@ import android.content.Context
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
+import net.aiscope.gdd_app.ui.sample_completion.metadata.MetadataFragment
+import net.aiscope.gdd_app.ui.sample_completion.preparation.PreparationFragment
+import net.aiscope.gdd_app.ui.sample_completion.quality.QualityFragment
 
 internal class FragmentAdapter(
     var context: Context,
