+++ app/src/main/res/layout/fragment_metadata.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.sample_completion.MetadataFragment">
+    tools:context=".ui.sample_completion.metadata.MetadataFragment">
 
     <ScrollView
         android:layout_width="match_parent"
