+++ app/src/main/res/layout/fragment_preparation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.sample_completion.PreparationFragment">
+    tools:context=".ui.sample_completion.preparation.PreparationFragment">
 
     <ScrollView
         android:layout_width="match_parent"
