+++ app/src/main/res/layout/fragment_quality.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.sample_completion.QualityFragment">
+    tools:context=".ui.sample_completion.quality.QualityFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
